+++ fpdf/enums.py
@@ -4,12 +4,6 @@ from sys import intern
 from .syntax import Name
 
 
-class DocumentState(IntEnum):  # internal/private enum
-    GENERATING = 0  # default state while building the PDF document
-    CLOSING = 1  # while calling .output() and converting the internal .pages dict to a .buffer
-    CLOSED = 2  # = .output() has been called once and has finished
-
-
 class SignatureFlag(IntEnum):
     SIGNATURES_EXIST = 1
     "If set, the document contains at least one signature field."
